+++ CHANGELOG.md
@@ -3,6 +3,8 @@
  - Added `byteAligned` combinator, which right pads a `Codec[A]` such that the binary size is always evenly divisible by 8.
  - Added `decodeValue` back to `Decoder`, `Codec`, and `GenCodec`.
  - Added `limitedSizeBits` and `limitedSizeBytes` combinators, which behave like `fixedSizeBits`/`fixedSizeBytes` but perform no padding.
+ - Added support for value codec framing to `DiscriminatorCodec` and auto-generated coproduct codecs. See `TlvExample` for usage.
+ - Added `fallback` and `discriminatorFallback` combinators.
 
 1.7.1
 =====