+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Added `limitedSizeBits` and `limitedSizeBytes` combinators, which behave like `fixedSizeBits`/`fixedSizeBytes` but perform no padding.
  - Added support for value codec framing to `DiscriminatorCodec` and auto-generated coproduct codecs. See `TlvExample` for usage.
  - Added `fallback` and `discriminatorFallback` combinators.
+ - Added `variableSizePrefixed{Bits,Bytes}` and `variableSizePrefixed{Bits,Bytes}Long` combinators for working with binary shapes like `size ++ prefix ++ value` where `size` encodes the length of `value`.
 
 1.7.1
 =====